Made LEN() return a Long instead of a Double.
git-svn-id: http://chianti.ucsd.edu/svn@20612 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ corelibs/trunk/equations/src/org/cytoscape/equations/builtins/Len.java
@@ -75,7 +75,7 @@ public class Len implements Function {
 	 */
 	public Object evaluateFunction(final Object[] args) throws IllegalArgumentException, ArithmeticException {
 		final String text = FunctionUtil.getArgAsString(args[0]);
-		return (Double)(double)text.length();
+		return (Long)(long)text.length();
 	}
 
 	/**
